Refactor current setup of definitions
